Move report_csp_violations out of globalscope
It doesn't use any internal members, so it can be in a separate
contained file.
