+++ components/script/script_runtime.rs
@@ -72,6 +72,7 @@ use crate::dom::bindings::reflector::{DomGlobal, DomObject};
 use crate::dom::bindings::root::trace_roots;
 use crate::dom::bindings::utils::DOM_CALLBACKS;
 use crate::dom::bindings::{principals, settings_stack};
+use crate::dom::csp::report_csp_violations;
 use crate::dom::event::{Event, EventBubbles, EventCancelable, EventStatus};
 use crate::dom::eventtarget::EventTarget;
 use crate::dom::globalscope::GlobalScope;
@@ -397,7 +398,7 @@ unsafe extern "C" fn content_security_policy_allows(
             RuntimeCode::WASM => csp_list.is_wasm_evaluation_allowed(),
         };
 
-        global.report_csp_violations(violations, None);
+        report_csp_violations(&global, violations, None);
         allowed = is_evaluation_allowed == CheckResult::Allowed;
     });
     allowed
